Resolve #84 - Move manpage to CLI project
This is really where this belongs.  Also, this makes it easier to add
build steps for it and to include it as content of the project (copied to
output dir on build).

By including the manpage as part of the project, including it in the
packaging project we get for free (mark it content and we're done).

* Update the HTML-creation script so that it exits with a zero return
  code if HTML tidy would have exited with a return of 1 (this is expected,
  it indicates there are warnings, which there will be).