WIP #258 - Finally nail the clone-element-with-source-info algorithm
This algorithm solves a problem I have been grappling on-and-off
for years now.  It's that I want to preserve the source information
when cloning an element, even though I'm creating new elements.

I also don't want the element's constructor logic to get in the way
and create more native elements.