Temporary commit to try and diagnose failing test
Oddly, on AppVeyor (only) this test fails for the test case
  GarbageMarkupInJavascript.html

But, when using the logs to try and determine what's wrong, it shows me that the
expected and actual renderings are identical.  Changing to use
Is.EqualTo might help determine why they're not being treated as equal by the
test constraint I wrote.