WIP #219 - Move a number of integration tests
This activates a number of them (many/most of which are failing).  Also,
a number have been moved to a new directory which indicates tests for
the i18n functionality.  These are not enabled yet because of:
  https://github.com/csf-dev/ZPT-Sharp/issues/4

Many of the failing tests will require further analysis and may even need
to be turned into tickets of their own.