Resolve #284 - Fix issues with on-error handler
It hadn't been actually populating the 'error' object, available to
consumers.  The integration test that pointed this out needed to be
slightly changed, though.   Its original source & expected renderings were
based upon implementation details of how Python/Zope works.  For a start
the .NET Exception object cannot be used with array notation (best we can
do is use ToString()) and secondly, the exception message differed, so the
actual output needed to be altered to match.