Resolve #261 - Create an XML resolver which uses embedded resources
The purpose of this is to permit the resolution of XHTML DTDs (and
supporting resources) without the use of network requests.  The
implementation locks down the resolution to only permit the resolution
of XHTML-related resources.

What this means for documents is that XHTML entities such as &nbsp; may be
used without causing errors in the XML reading processs.  The DTDs and
supporting resources are provided within this assembly as embedded
resources and returned without the use of network requests.