Trivial - update multi-targeting
There's no need for explicit net47 or net48 packages, because these
targets are sufficiently covered by netstandard2.0.

There is a reason to explicitly include net461, though, in line
with guidance at
  https://docs.microsoft.com/en-us/dotnet/standard/library-guidance/cross-platform-targeting#multi-targeting

In addition, I want to support net46 everywhere and additionally down
as far as net45 in most libraries where I can.  The reason for separating
net45 and net452 is because there were some breaking changes in net452,
which could plausibly mean that separate versions are useful.