Resolve #259 - Prove that ViewBag now works in path expressions
This sample code change to the example proves that the ViewBag now works.
This has actually been fixed for some time in ZptSharp 2.x by the handling
of Dynamic Types in path expressions.

See ZptSharp.Impl/Expressions/PathExpressions/DynamicObjectValueProvider.cs
for the logic which actually makes this work.  It has full test coverage.

This commit will be reverted-out of the repository though, becaise I do not
want to make it a permanent part of the solution.  I just wanted to prove
that it does in fact work.