WIP #265 - Initial work on self-hosting
I found that this was somewhat required in order to implement
MVC ViewEngines properly, particularly in MVC Core.

The reason is that the view-engine is expected to be fully
created and initialised whilst registering services with the
service collection.  However, ZptSharp requires quite a bit of
configuration after the service collection is created.  Thus,
this provides a way to use a small, self-contained service
collection to initalise the ZptSharp environment.