Resolve #265 - Complete self-hoster
This also includes some rework to the integration tests, so that they now make
use of the self-hoster instead of setting up the environment themselves.

I have also had to drop support for .NET Framework versions lower
than 4.6.1 in this work.  Unfortunately a breaking change in the MS
dependency injections library means that I cannot use version 1.x with
this project.