WIP #263 - Add MVC Core ViewEngine
This involves rather a lot of changes, because I have now
integrated both MVC5 and MVC Core view engines into (largely)
common source code.

There are a couple of files with rather a lot of compiler flags in them
and perhaps I can improve upon that later.

Also, test coverage (and an example MVC Core app) need to be written.