WIP #263 - Rework some of the MVC types
This is mainly to avoid reliance on compiler flags, which make the
types more difficult to read & understand.

The technique used here is mainly the use of a frmaeowkr-neutral
base class and also splitting some framework-sensitive operations
away into separate services, which may be built in the
framework-specific projects.

The result is that most of the code for the MVC projects:
* Is now unit-testable
* Does not rely upon compiler flags anymore