Improve build
It occurred to me that some files are not being reported-upon
by SonarCloud at the moment.  That's because they aren't being examined,
because they are never built.  When running only tests, the only
projects which are built are those referenced by tests.

An extra build at the beginning is otherwise not needed, but it does
mean that everything is analysed.