WIP #271 - Another attempt to fix tests
It seems that the globbing package is hard-coded to use
forward-slashes in its results, regardless of whether the
OS uses forward, or back-slashes.  That means that I have
to post-process the results in order to replace them.

That should give me a path formatted in the correct way for
the environment.