WIP #264 - Reduce package size
The main thrust of this work is ensuring that the net461 version of
the package does not require any references to netstandard.
A few of the dependencies I had been using required netstandard,
particularly the Microsoft.Extensions ones.  By upgrading these to
v5.0.0, I am now using versions which have net461 builds and do not
require netstandard dependencies.

Also, AngleSharp had been attempting to use the netstandard
version instead of the net461 version.  I had to manually change that.