WIP #270 - Refactoring for load expressions
This quite a significant change.  Mainly it removes the layer
previously called the "request renderer" and removes the concept
of a rendering request altogether.  Really, it's just a model, stream
and source info anyway.  That is very quickly replaced by document
and model.

The idea is that it will now be easier to render an arbitrary expression
context without the need to create a new rendering request.